pyproject.toml
[project]
name = "cdc-pipeline"
version = "0.1.0"
description = "Phase 1: Coin Market Cap -> Postgres -> Debezium -> Kafka -> S3"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32",
    "psycopg[binary,pool]>=3.2",
    "python-dotenv>=1.0",
    "tenacity>=9.0",
]

[project.scripts]
cdc-pipeline = "cdc_pipeline.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cdc_pipeline"]
